feat: Configure API call options
Added the ability to select and configure LLMs, including options such as temperature and prompt customization. Implementation-wise, it is all stored and persisted in the `AIState` object within the AI context and saved in the DB.

LLMs can come remotely (hard-coded in`models.ts`) or created locally (in local storage) using a form component (`model-form.tsx`). They and displayed in a nice static model card